begin sound implementation, correct textbox offset

+++ TextBox.h
@@ -9,7 +9,7 @@ namespace penguine
 	class TextBox: public Component
 	{
 	protected:
-		sf::Vector2i* m_Offset;
+		sf::Vector2f* m_Offset;
 		sf::Color m_FillColor;
 
 		sf::Vector2u* m_Dimensions;
